AYo kerja guys mantap guys

+++ core_functionality.py
@@ -18,7 +18,7 @@ def read_n_dict(file_path, MAIN_KEY, groupables, RANKED_FIELD, RANK_STORAGE):
             name = row_.pop(MAIN_KEY) # the Name field becomes the keys
             try :
                 row_[RANKED_FIELD] = float(row_[RANKED_FIELD]) # convert these as numerical value
-                row_[RANK_STORAGE] = float(row_[RANK_STORAGE]) # convert these as numerical value
+                row_[RANK_STORAGE] = int(row_[RANK_STORAGE]) # convert these as numerical value
             except ValueError: continue # skip invalid data
 
             main_dict[name] = row_ # main contains all the fields as value of each Name as key
